Fix so step scans with 1 um step will export
With fly scans, the index 5 corresponds to number of steps. If this is
one, it is likely an alignment scan so it should be disregarded.

If this is a step scan, the index 5 corresponds to the step size so a
value of 1 means 1 um steps and it should still be exported.

This fix checks for the scan type and the value of the index to decide
if it should be ignored.